backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
apscheduler==3.10.4
aiosqlite==0.20.0
pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.8.2
aiohttp==3.10.5
tvdatafeed==2.3.0
fredapi==0.5.2
httpx==0.27.2
